+++ rpn.h
@@ -5,7 +5,7 @@
 
 typedef struct {
 	int sp;
-	float elems[RPN_STACK_SIZE];
+	double elems[RPN_STACK_SIZE];
 } RPNStack;
 
 enum {
@@ -19,7 +19,7 @@ enum {
 RPNStack *rpn_stack_init(RPNStack *stack);
 int rpn_stack_drop(RPNStack *stack);
 int rpn_stack_dup(RPNStack *stack);
-int rpn_stack_peek(float *dest, RPNStack stack);
+int rpn_stack_peek(double *dest, RPNStack stack);
 int rpn_stack_swap(RPNStack *stack);
-int rpn_calc(float *dest, const char *expr, RPNStack *stack);
+int rpn_calc(double *dest, const char *expr, RPNStack *stack);
 const char *rpn_strerr(int rpnerr);
